move bounded buffer code to separate module
