+++ bd-logger/src/pre_config_buffer_test.rs
@@ -5,7 +5,7 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use crate::memory_bound::MemorySized;
+use crate::bounded_buffer::MemorySized;
 use crate::pre_config_buffer::{self, PreConfigBuffer};
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
